+++ src/test/java/uk/gov/hmcts/darts/arm/mapper/impl/MediaArchiveRecordMapperImplTest.java
@@ -175,8 +175,8 @@ class MediaArchiveRecordMapperImplTest {
         assertEquals("2025-01-23T10:30:00.000Z", metadata.getBf011());
         assertEquals(mediaEntity.getId(), metadata.getBf012());
         assertEquals(mediaEntity.getId(), metadata.getBf013());
-        assertEquals(mediaEntity.getChannel(), metadata.getBf014());
-        assertEquals(mediaEntity.getTotalChannels(), metadata.getBf015());
+        assertEquals(mediaEntity.getChannel().longValue(), metadata.getBf014());
+        assertEquals(mediaEntity.getTotalChannels().longValue(), metadata.getBf015());
         assertNull(metadata.getBf016());
         assertEquals("2025-01-23T17:30:00.000Z", metadata.getBf017());
         assertNull(metadata.getBf018());
@@ -198,8 +198,8 @@ class MediaArchiveRecordMapperImplTest {
         assertNotNull(metadata.getBf011());
         assertEquals(mediaEntity.getId(), metadata.getBf012());
         assertEquals(mediaEntity.getId(), metadata.getBf013());
-        assertEquals(mediaEntity.getChannel(), metadata.getBf014());
-        assertEquals(mediaEntity.getTotalChannels(), metadata.getBf015());
+        assertEquals(mediaEntity.getChannel().longValue(), metadata.getBf014());
+        assertEquals(mediaEntity.getTotalChannels().longValue(), metadata.getBf015());
         assertNull(metadata.getBf016());
         assertEquals("2025-01-23T10:30:00.000Z", metadata.getBf017());
         assertEquals("2025-01-23T17:30:00.000Z", metadata.getBf018());
